LayoutTests/http/tests/inspector/network/copy-as-fetch-expected.txt

Tests that "Copy as fetch" returns valid JS.


== Running test suite: WI.Resource.prototype.generateFetchCode
-- Running test case: WI.Resource.prototype.generateFetchCode.SimpleURL
fetch("http://127.0.0.1:8000/inspector/network/resources/url?query=true", {
    "cache": "default",
    "credentials": "omit",
    "headers": {
        "Accept": "*/*",
        "User-Agent": <filtered>
    },
    "method": "GET",
    "mode": "cors",
    "redirect": "follow",
    "referrer": "http://127.0.0.1:8000/inspector/network/copy-as-fetch.html"
})

-- Running test case: WI.Resource.prototype.generateFetchCode.SpecialURL
fetch("http://127.0.0.1:8000/inspector/network/resources/url'with$special%7B1..20%7Dchars[]%20.html", {
    "cache": "default",
    "credentials": "omit",
    "headers": {
        "Accept": "*/*",
        "User-Agent": <filtered>
    },
    "method": "GET",
    "mode": "cors",
    "redirect": "follow",
    "referrer": "http://127.0.0.1:8000/inspector/network/copy-as-fetch.html"
})

-- Running test case: WI.Resource.prototype.generateFetchCode.SpecialHeaders
fetch("http://127.0.0.1:8000/inspector/network/resources/url", {
    "cache": "default",
    "credentials": "omit",
    "headers": {
        "Accept": "*/*",
        "User-Agent": <filtered>,
        "X-Custom1": "test1",
        "X-Custom2'%": "'Test''1\\'2",
        "X-Custom3": "'${PWD}"
    },
    "method": "GET",
    "mode": "cors",
    "redirect": "follow",
    "referrer": "http://127.0.0.1:8000/inspector/network/copy-as-fetch.html"
})

-- Running test case: WI.Resource.prototype.generateFetchCode.URLUTF8
fetch("http://127.0.0.1:8000/inspector/network/resources/url?utf8=%F0%9F%91%8D", {
    "cache": "default",
    "credentials": "omit",
    "headers": {
        "Accept": "*/*",
        "User-Agent": <filtered>
    },
    "method": "GET",
    "mode": "cors",
    "redirect": "follow",
    "referrer": "http://127.0.0.1:8000/inspector/network/copy-as-fetch.html"
})

-- Running test case: WI.Resource.prototype.generateFetchCode.POSTRequestURLEncodedData
fetch("http://127.0.0.1:8000/inspector/network/resources/url", {
    "body": "lorem=ipsum&$dolor='sit'&amet={1..20}",
    "cache": "default",
    "credentials": "omit",
    "headers": {
        "Accept": "*/*",
        "Content-Type": "application/x-www-form-urlencoded",
        "User-Agent": <filtered>
    },
    "method": "POST",
    "mode": "cors",
    "redirect": "follow",
    "referrer": "http://127.0.0.1:8000/inspector/network/copy-as-fetch.html"
})

-- Running test case: WI.Resource.prototype.generateFetchCode.POSTRequestURLUTF8
fetch("http://127.0.0.1:8000/inspector/network/resources/url?utf8=%F0%9F%91%8D", {
    "body": "🌨=⛄️",
    "cache": "default",
    "credentials": "omit",
    "headers": {
        "Accept": "*/*",
        "Content-Type": "application/x-www-form-urlencoded",
        "User-Agent": <filtered>
    },
    "method": "POST",
    "mode": "cors",
    "redirect": "follow",
    "referrer": "http://127.0.0.1:8000/inspector/network/copy-as-fetch.html"
})

-- Running test case: WI.Resource.prototype.generateFetchCode.PUTRequestWithJSON
fetch("http://127.0.0.1:8000/inspector/network/resources/url", {
    "body": "{\"update\":\"now\"}",
    "cache": "default",
    "credentials": "omit",
    "headers": {
        "Accept": "*/*",
        "Content-Type": "application/json",
        "User-Agent": <filtered>
    },
    "method": "PUT",
    "mode": "cors",
    "redirect": "follow",
    "referrer": "http://127.0.0.1:8000/inspector/network/copy-as-fetch.html"
})
